Web: Adds routing handling for SPA with GitHub Pages
Adds handling for hosting a SPA with multiple "pages" (such as this
application) on GitHub Pages by copying the built `index.html` file to
`404.html`.

The issue is that GitHub Pages doesn't know about the routing of the SPA
and will give a 404 error when navigating directly to a non-root path.
This makes it so that the 404 page is just the app itself.